Replace t1-volume by successive calls of t1-volume* sub-pipelines
