+++ clinica/iotools/utils/pipeline_handling.py
@@ -195,7 +195,7 @@ def t1_volume_pipeline(caps_dir, df, **kwargs):
     Returns:
         final_df: a DataFrame containing the information of the bids and the pipeline
     """
-    from ...pipelines.t1_volume_new_template.t1_volume_new_template_cli import T1VolumeNewTemplateCLI
+    from ...pipelines.t1_volume.t1_volume_cli import T1VolumeNewTemplateCLI
     instance = T1VolumeNewTemplateCLI()
     pipeline_name = instance.name
     del instance
