+++ clinica/pipelines/t1_volume/info.json
@@ -1,5 +1,5 @@
 {
-    "id": "aramislab/t1-volume-new-template",
+    "id": "aramislab/t1-volume",
     "author": "Jorge Samper-Gonzalez; Simona Bottani",
     "version": "0.1.0",
     "dependencies": [
